feat: DT-466 Support `dev.email` trace tag
[DT-466](https://outreach-io.atlassian.net/browse/DT-466)

Updates the `trace.yaml` config file to support a new section,
`GlobalTags`.  The idea here is that `GlobalTags` will allow us to
specify in configs certain tags we would like to add to all traces.  The
`MarshalLog` implementation for the `GlobalTags` config struct will
handle converions to standardized tag names.

We start with just one attribute: `DevEmail`/`dev.email`.  This will be
injected into pods via bootstrap so that apps running in the local dev
environment will be tagged with the developer's email, making the
Honeycomb `dev` dataset much easier to navigate.

This is a forward-port of
https://github.com/getoutreach/go-outreach/pull/165.