feat: refactor pkg/trace
No changes in any exported functionality, though we can do that in followups.

The main goal of this change is to come up with a clean shape for tracers and separate the honeycomb/call-tracking/metrics/logging dependencies.

1. Adds internal/tracer with provider setup and pure honeycomb and call-tracker tracers.
2. Adds internal/tracelog with a singleton tracer that also adds the metrics + logging tracers.
3: Updates pkg/trace to just forward all requests to the singleton in internal/tracelog.

I plan to do another cleanup:

1. Move internal/call directy into internal/tracer and remove its dependency on /pkg/metrics
2. Move metrics dependency over to internal/tracelog direclty.

Note that most of the packages are in /internal as it will help move things around safely without backwards compatiblity issues.