feat: add experimental tracelog provider
This change restructures pkg/trace to allow adding an experimental trace logger which cna log extra info in trace.StartCall/trace.EndCall.

Note that the experimental tracelogger itself is in x/tracelog and as such we can iterate on it without backwards compatibility issues.

Also note while this change adds trace.AddProvider, there is no major refactoring of trace itself as that is deferred to separate PR.