feat(trace): SpanStartOption and WithLinks (#220)
<!--
  !!!! README !!!! Please fill this out.

  Please follow the PR naming conventions: 

https://outreach-io.atlassian.net/wiki/spaces/EN/pages/1902444645/Conventional+Commits
-->


<!-- A short description of what your PR does and what it solves. -->
## What this PR does / why we need it
Clerk needs to link producer spans with consumer ones cross VM
boundaries. In order to do so, we need gobox's trace support to start a
span with a link. Clerk already propagates traceHeaders as part of the
event, so this PR is to take the same headers (map[string][]string -
same as existing To/FromHeaders methods) and generate a link out of them
+ attach the link as a new StartSpanOption. StartSpanOption abstracts
out otel implementation ...

<!--- Block(jiraPrefix) --->

## Jira ID

[CDT-102]

<!--- EndBlock(jiraPrefix) --->

<!-- Notes that may be helpful for anyone reviewing this PR -->
## Notes for your reviewers



<!--- Block(custom) -->

<!--- EndBlock(custom) -->


[CDT-102]:
https://outreach-io.atlassian.net/browse/CDT-102?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ