fix(trace): intro Link with Trace/Span ID funcs and and export trace.SpanID (#221)
<!--
  !!!! README !!!! Please fill this out.

  Please follow the PR naming conventions: 

https://outreach-io.atlassian.net/wiki/spaces/EN/pages/1902444645/Conventional+Commits
-->


<!-- A short description of what your PR does and what it solves. -->
## What this PR does / why we need it
This PR exposes SpanID. It also exports the former linkOption as a Link
struct to allow callers access the trace/span IDs directly from the
link.

Clerk needs to link a trace to process batch of clerk events to their
producer's traces. If producer sent them in a batch, there is a good
chance that there will be large overlaps between producer and consumer
batches with span IDs, resulting in multiple links with same trace/span
ID.

I checked otel's code - it does not deduplicate links. If the # of links
is over limit (default 128), it will simply evict others.

https://github.com/open-telemetry/opentelemetry-go/blob/55b49c407e07dc13bdba245c2c13935a37bad971/sdk/trace/tracer.go#L144.
SpanID is needed to deduplicate the links.


<!--- Block(jiraPrefix) --->

## Jira ID

[CDT-102]

<!--- EndBlock(jiraPrefix) --->

<!-- Notes that may be helpful for anyone reviewing this PR -->
## Notes for your reviewers



<!--- Block(custom) -->

<!--- EndBlock(custom) -->


[CDT-102]:
https://outreach-io.atlassian.net/browse/CDT-102?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ