DRAFT: Proper types for OTel span attrs
A proof-of-concept that attempts to use the right attrbute type for the
given input.  As of right now, everything is being reported up as a
string, and HC sees it as a string, which means we don't get support for
`SUM` or `AVG` or other numeric-only functions on numeric data.

We can always fix this manually in the UX, but ideally we shouldn't need
to.