feat: STAR-760 Support typed OTel span attrs (#248)
[STAR-760](https://outreach-io.atlassian.net/browse/STAR-760)

Updates the OTel `MarshalLog` serialization with special handling for
most Go scalar types. Previously, anything that ended up as a "value" in
`MarhsalLog` and which was not itself "Marshalable" would be
stringified. The new code will attempt to preserve types.

This fine idea breaks down in a few places at the implementation detail
level. Since OTel only supports "int64" and not "uint64" we have no
choice but to continue stringify-ing `uint64` and `uint` values. Also,
it turns out there's a lot of boilerplate code involved handling all the
many Golang integer types.

This commit notably (and perhaps controversially) adds special support
for `time.Time`. The OTel specs are a mess, so there's no explicit
"time" attribute type or any solid conventions on how times should be
represented. It seems good to abstract this serialization decision away
from the folks who are implementating Marshalers, since the "right" time
format might vary based on context or which trace provider we're using.
This commit's implementation is to serialize it as an RFC3389 string.

Updates a bunch of tests and test helpers to account for these new
behaviors. Adds a bunch of new tests to exhaustively test the new tag
serialization method.