fix(cli): small fixes to logic in CLI package(s) (#255)
This PR contains fixes to a few CLI-oriented packages. Primarily it
contains a fix for the `resolver` package.

Resolver Fix:

This fix ensures that when consider a branches (with `AllowBranches`)
that we do not insert a mutable version for that branch on to the
channel. The problem with doing so is that a mutable version is always
ranked higher than tags, this is problematic for branches that _only_
release and support tags on the branch (e.g. `rc` vs `unstable`).

We now filter out the mutable insertion after converting references into
their channel/version pair, if we have a tag then we remove the latest
version (if it's mutable). It will always be the latest version because
of the sorting logic (which broke this library) in the previous PR.

sshhelper Fix:

This is a small fix to the logic to support read-only ssh-agents (like
1Password) and to allow users to ignore some previously fatal errors
that may be "solvable" with their ssh-agent. I needed this for most of
our tooling (I use the 1Password ssh-agent)