feat(trace): set error status when there are errors (#258)
Set otel status; add trace sugar

Errors are annoying right now.

You have to convert an `error` into a log.Marshaler, usually through
events.ErrorInfo, (which isn't itself an error, so you have to return
your original error from your function), and then, for your trouble, you
don't even get the error status set correctly on your span.

I am aware the call framework exists, and likely does nice things on the
span when used, but for library authors (e.g. QSS & smartstore) it
doesn't make sense to either assume we are in or start a call, most of
the time, but we still want to know when our infrastructure errors in
traces.

This commit adds an Error convenience function, and special handling in
otel.go for both `error` and `events.ErrorInfo` objects so the status is
set. This should improve the utility of honeycomb, and other tools that
assume you'll use the OTEL spec's `code` field. The commit should be
considered a proposal until discussion with tooling is resolved, and
tests added.

<!--
  !!!! README !!!! Please fill this out.

  Please follow the PR naming conventions: 

https://outreach-io.atlassian.net/wiki/spaces/EN/pages/1902444645/Conventional+Commits
-->


<!-- A short description of what your PR does and what it solves. -->
## What this PR does / why we need it



<!-- <<Stencil::Block(jiraPrefix)>> -->

## Jira ID

[XX-XX]

<!-- <</Stencil::Block>> -->

<!-- Notes that may be helpful for anyone reviewing this PR -->
## Notes for your reviewers



<!-- <<Stencil::Block(custom)>> -->

<!-- <</Stencil::Block>> -->

---------

Co-authored-by: CircleCI <circleci@outreach.io>