fix(aws): properly auth when expires is empty, add/fix some tests (#287)
<!--
  !!!! README !!!! Please fill this out.

  Please follow the PR naming conventions: 

https://outreach-io.atlassian.net/wiki/spaces/EN/pages/1902444645/Conventional+Commits
-->


<!-- A short description of what your PR does and what it solves. -->
## What this PR does / why we need it

If `~/.aws/credentials` existed (or maybe if it didn't even exist at
all) we'd not refresh the auth properly, causing the user to have to run
`saml2aws` at least once. This fixes that and adds a bunch of tests to
ensure we always... do that. Also fixed an otel logfile test and some
spelling

<!-- <<Stencil::Block(jiraPrefix)>> -->

## Jira ID

[XX-XX]

<!-- <</Stencil::Block>> -->

<!-- Notes that may be helpful for anyone reviewing this PR -->
## Notes for your reviewers



<!-- <<Stencil::Block(custom)>> -->

<!-- <</Stencil::Block>> -->