feat(trace): Support disabling logging for trace call methods (#300)
Adds a `CallOptions` type that can be used for specific options
for tracing calls. Ideally, this would be limited to artifacts of traces
(e.g., logs). The first, and only, option implemented is
`DisableInfoLogging`. This defaults to `false`. When set to `true`, info
logs are not generated when a call ends.'

I also cleaned up some tests to no longer be attached to the test suite
struct invoked by `shuffler.Run()`. This helps us when we fully remove
the shuffler in favor of using standard go testing.

Updated the `gopls` build flags to not include `or_e2e` for invocation
of tests through VSCode and compilation. This doesn't really matter
except for some packages, ideally we'll remove these tags entirely in
the future when we stop using build tags to handle configuration loading
logic (e.g., move entirely to devspace).

---------

Co-authored-by: Jesse Kinkead <jesse.kinkead@outreach.io>