feat: Updating otel to latest versions, fixing unit tests (#322)
To fix some dependency issues in vault-client, I had to walk back up the
chain to gobox to update to the latest otel libraries. This caused
several changes in attributes that the library is sending out, changing
the unit test expected results. It's unclear if this is idiomatic
changes or what, so I'm not sure how to review this.