feat: Deprecate async/pool in favor of github.com/sourcegraph/conc/pool
Deprecates the major methods in `async/pool` in favor of using
`github.com/sourcegraph/conc/pool` instead. The `async/pool` has had
usability issues for a long time now, namely hard to use `Wait()` as
well as almost impossible to get errors out of. The `conc/pool` library
is a much better, more actively maintained libary.

Included is a new `README.md` with instructions on how to migrate all
usecases I could discern from across the `getoutreach` organization.

The main change here is the removal of bufferred pools of work items
that could be dynamically resized as well as rejected if there was no
room. I could only find one repo, `mailroomapi`, that used this
functionality. It is likely that the block until accepted behaviour will
be more acceptable to this service, but also I don't feel that warrants
the additional overhead/code duplication that we own here today.