fix(updater): migrate to fork of go-update (#346)
`go-update` hasn't been updated in 8 years and the company backing it
has gone under. Moving to the fork maintained by fynelabs. Given the
library is meant to provide an automated updater interface, their
breaking changes don't apply to our _very_ limited use of it (we just
use it for cross-platform replacements of the running binary).

Worked when testing locally.