feat: add capability to have a status code for client canceled. (#349)
<!--
  !!!! README !!!! Please fill this out.

  Please follow the PR naming conventions: 

https://outreach-io.atlassian.net/wiki/spaces/EN/pages/1902444645/Conventional+Commits
-->


<!-- A short description of what your PR does and what it solves. -->
## What this PR does / why we need it
This happens when a server already cancels a request. This will enable
better metrics tracking. There were cases where due to a context being
canceled an InternalServerError returned. This would cause alerts to
trigger when in fact the only issue was that the context was canceled.


<!-- <<Stencil::Block(jiraPrefix)>> -->

## Jira ID

[STAR-1788]

<!-- <</Stencil::Block>> -->

<!-- Notes that may be helpful for anyone reviewing this PR -->
## Notes for your reviewers
This was an issue that was happening for marathon. By the time the
server goes to make the request the context has been canceled. At this
point its best to simply return "ClientCanceled"

More discussion came from this slack thread.
https://outreach-hq.slack.com/archives/CN9MU7GLW/p1691689393028399

This PR is simply the first part of what was already discussed.
<!-- <<Stencil::Block(custom)>> -->

<!-- <</Stencil::Block>> -->


[STAR-1788]:
https://outreach-io.atlassian.net/browse/STAR-1788?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: Jared Allard <jaredallard@users.noreply.github.com>