feat: DT-4029: Disable call info logging default (#365)
[DT-4029](https://outreach-io.atlassian.net/browse/DT-4029)

Disables the info logs that would previously occur as part of a
`trace.StartCall`'s succesful completion.  Calls that end in errors have
been and will continue to be logged.

Adds a boolean config option to `trace.yaml` so that services can opt
back in to the old behavior.  We don't want services to fear upgrading
to this version of gobox; this flag will let them upgrade without
experiencing an unpleasant behavior change.

Adds a new call option to explicitly enable info logs for particular
instances of `trace.StartCall`.  We hope that folks that rely on these
logs will rely on this more targeted approach to meet their needs,
rather than using the service-wide `trace.yaml` opt-out.

If you would like to keep your service on the more verbose behavior,
add the following snippet to your service's `.override.jsonnet`:
```
{
  trace_configmap+: {
    data_+:: {
      LogCallsByDefault: true,
    }
  }
}
```

[DT-4029]:
https://outreach-io.atlassian.net/browse/DT-4029?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ