high level impl of olog with slog backend and per-module log levels
PoC implementation of a logger using `slog.Logger` as the primary type
with wrappers around it. Adds a global registry used for determining the
log level of a logger based on the module. Eventually, `New()` will
associate loggers with the current package and module name.

LogLevels can be set based on module or package name using the
`(*levelRegistry).Set` function.

While it would be more performant to store the `leveler` instances in
the registry and update those on `Set`, I feel that would complicate the
code. I also feel that a should a logger become a bottleneck in the
execution of an application, it should be removed in favor of tracing or
metrics gathering.