feat(ometrics): adds the ometrics package
This adds a `ometrics` package (as a sub-module, more on that later)
aimed at configuring the global otel.MeterProvider. This allows
consumers to instrument metrics through the standard otel metrics
package. The ideal usage is below:

```go
ometrics.InitializeMeterProvider(ctx, ExporterTypePrometheus)

var meter otel.GetMeterProvider().Meter("packageName")

func doAThing() {
  counter, err := meter.Float64Counter("my_counter", api.WithDescription("counts a thing"))
  if err != nil {}
  counter.Add(ctx, 5)
}
```

**Note**: Yes, there is not a notion of package global metrics like
Prometheus had. Metrics have to be instrumented in a function and then
passed along to whatever consumes them. That was a design decision by
otel maintainers.

There's some open questions in comments in the code that I was unable to
get to as well as future direction. For example, to fully support the
collector exporter (push) there needs to be configuration instrumented.
It's up to the future implementor to decide how that should be done.
Though, I would suggest not utilizing environment variables for it :)

This is added a sub-module because the otel dependencies have been
_very_ painful in the past, so I'd like to avoid proliferating them to
every project that consumes `gobox`. Otherwise, just sourcing `pkg/app`
can break an application today.