fix(trace): Support parent-based trace sampling (DT-4122) (#417)
[DT-4122](https://outreach-io.atlassian.net/browse/DT-4122)

Updates the gobox trace library to support OpenTelemetry parent-based
sampling.  Now, when a Go service receives a traceparent header that
includes the "sampled" bit, indicating that the parent sampled this
trace, then the Go service will also sample this trace.

This contrasts with the previous way of doing things, inherited from
Honeycomb beeline libraries, where each service independently computed
a deterministic hash of the trace ID and used that to make independent
but convergent sampling decisions.

The motivation for this change is the fact that Gazelle/Giraffe, and
later the flagship, have hopped on board this more standard parent-based
sampling convention.  We've been seeing broken traces as the Go services
lag behind and make different sampling decisions based on different
algorithms.

This change is complicated by the fact that we want to retain support
for Honeycomb's "sample rate" attribute, which auto-adjusts the counts
and sums in their UI to account for sampling.  There is no finalized
OpenTelemetry standard for this.  We implement our own solution, using
the (cross-service spanning) "tracestate" attribute to carry information
about the sample rate that was used to make the sampling decision.  We
don't expect cross-service propagation of this attribute to work *yet*,
but the pieces are there that we could attempt this in the future.

Breaks the existing sampler into its component parts, defining new and
separate samplers for the "honeycomb deterministic sampling" and
"x-force-trace" and "hack to attach sample rate attributes".  These
samplers are more unit-testable than the all-in-one samplers that
preceded them.  We also provide one `defaultSampler` that ties together
all these pieces, and we make it the one and only option for gobox OTel
tracing for now.

[DT-4122]:
https://outreach-io.atlassian.net/browse/DT-4122?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ