feat(olog): add charm-style formatter for logrus (#525)
Add a charm-style formatter for logrus to have consistent logging
formats.

`log/slog` has been widely accepted by the community as the new standard
for logging in Go. This olog package is our implementation of it which
uses the charm logger. However, we are currently using logrus in many
places for CLI applications. In order to have consistency in our log
formats, I decided to create a custom logrus formatter which matches the
text format of the charm logger.