feat: upgrade to Go 1.22.6 (#589)
## What this PR does / why we need it

It's significantly more complicated to upgrade Go in this repo since
it's effectively a monorepo now. This means that restenciling with Go
minor version bumps will not work without PRs like this.

I also had to update the `golang.org/x/tools/cmd/stringer` version as
using an ancient version broke with Go 1.22.

## Notes for reviewers

A restencil was needed due to various incompatibilities. This was added
to a separate commmit.