feat!(rm-cause): causes are harmful
This code is demented. It produces messages like the folliowing
(irrelvant fields redacted):

```json
{
	"content": {
		"attributes": {
			"error": {
				"kind": "error",
				"cause": {
					"kind": "cause",
					"cause": {
						"kind": "cause",
						"cause": {
							"kind": "cause",
							"message": "10.96.232.125:5432 (app2a-callstatemanager-xpmuvqngpcd.cluster-ro-cntwmzilna0e.us-east-1.rds.amazonaws.com): failed to write startup message: context canceled"
						},
						"message": "failed to connect to `user=readwrite database=callstatemanager`"
					},
					"message": "failed to connect to `user=readwrite database=callstatemanager`: 10.96.232.125:5432 (app2a-callstatemanager-xpmuvqngpcd.cluster-ro-cntwmzilna0e.us-east-1.rds.amazonaws.com): failed to write startup message: context canceled getting connection to host app2a-callstatemanager-xpmuvqngpcd.cluster-ro-cntwmzilna0e.us-east-1.rds.amazonaws.com:5432 for org a5fc3e3c-b376-4327-9d80-3e29fe62e6e5"
				},
				"error": "failed to connect to `user=readwrite database=callstatemanager`: 10.96.232.125:5432 (app2a-callstatemanager-xpmuvqngpcd.cluster-ro-cntwmzilna0e.us-east-1.rds.amazonaws.com): failed to write startup message: context canceled getting connection to host app2a-callstatemanager-xpmuvqngpcd.cluster-ro-cntwmzilna0e.us-east-1.rds.amazonaws.com:5432 for org a5fc3e3c-b376-4327-9d80-3e29fe62e6e5",
				"message": ""
			},
		}
	}
}
```

In order to query this, you end up writing statements like
`error.cause.cause.cause.cause.message`

There's no point in that. All of our tools also support "in" queries,
so you can just check `if "context cancelled" in error.error`
(error.error) is also semantically problematic, buit whatvs.

All this does is inflate our log volume and increase meaninglessly the
number of columns in honeycomb queries.

Remove it and profit.

NB: this is a scorched earch approach. Anyone updating would be
immediately su bject to this new behavior, and it might break alerting
if someone was specifically looking for the 5th cause or something goofy
like that.