feat(slog): make gobox a facade atop slog
slog is the not-so-new structured logger in the standard lib for golang

using it is beneficial. In particular, we can use https://pkg.go.dev/go.opentelemetry.io/contrib/bridges/otelslog
to get the real trace into the log message, and then leverage that
elsewhere to e.g. [ensure all logs for a trace are captured][prob-sampl-proc]

[prob-sample-proc]: https://github.com/open-telemetry/opentelemetry-collector-contrib/blob/main/processor/probabilisticsamplerprocessor/README.md#logs-specific-configuration