feat(cli/updater): skip updating if installed via external method (#663)
## What this PR does / why we need it

Adds `skipPaths` to the updater config so that if the binaries are
installed in a subpath defined in the list of subpaths, then it will
disable auto-updating.

For example, if `stencil` was installed via Homebrew, it would be
installed somewhere in `/usr/local/Cellar`. If it was installed via
`mise`, it would be installed in a subdirectory of
`$HOME/.local/share/mise/installs/`. `skipPaths` would look like:

```yaml
global:
  skipPaths:
    - /usr/local/Cellar/
    - /.local/share/mise/installs/
```

Note that there is not currently support for per-repo `skipPaths`, even
though it exists on the config object.

## Jira ID

[DT-4756]

## Notes for your reviewers

In order to be able to write the paths automatically to the updater
config file, I made the Config struct public.

[DT-4756]:
https://outreach-io.atlassian.net/browse/DT-4756?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ