feat(trace): fix trace.Error and support sending events directly. (#695)
events are the way your supposed to send information about points in
time during your span. See
https://opentelemetry.io/docs/specs/otel/trace/api/#add-events This PR
adds support for them, because I wanted to use them in polaroid to
report progress on checking an SLo without creating tiny lil nanosecond
duration spans.

trace.Error was supposed to do the proper error reporting, but it did so
in a bit of a convoluted way and the testing utility actually hid the
necessary component from the developer, so I exposed that utility, wrote
a better test, and then updated trace.Error to actually consistently
report errors.

<!--
  !!!! README !!!! Please fill this out.

  Please follow conventional commit naming conventions:

  https://www.conventionalcommits.org/en/v1.0.0/#summary
-->

Please read [CONTRIBUTING.md](CONTRIBUTING.md) for additional
information on contributing to this repository!

<!-- A short description of what your PR does and what it solves. -->
## What this PR does / why we need it


<!-- <<Stencil::Block(jiraPrefix)>> -->

## Jira ID

[XX-XX]

<!-- <</Stencil::Block>> -->

<!-- Notes that may be helpful for anyone reviewing this PR -->
## Notes for your reviewers

<!-- <<Stencil::Block(custom)>> -->

<!-- <</Stencil::Block>> -->